requirements.txt
pyTelegramBotAPI
yt-dlp
cloudinary
python-dotenv
flask
gunicorn